requirements.txt
pycryptodome
httpx
tqdm
UnityPy
argparse